+++ app/Http/Controllers/Business/StockController.php
@@ -7,14 +7,21 @@ use App\Models\Stock;
 use App\Models\Product;
 use App\Models\Business;
 use Illuminate\Http\Request;
-use App\Http\Controllers\Controller;
+use App\Helpers\BusinessUserHelper;
 
+use App\Http\Controllers\Controller;
+use Illuminate\Support\Facades\Auth;
 use Symfony\Component\HttpFoundation\Response;
 
 class StockController extends Controller
 {
     public function index(Business $business, Request $request)
     {   
+        $businessUser = BusinessUserHelper::index($business['id'], Auth::user()['id']);
+        
+        if (!$businessUser) {
+            return abort(403);
+        } 
         $search = $request['search'];
         
         $stocks = Stock::whereHas('product', function($query) use ($search, $business){
@@ -32,6 +39,11 @@ class StockController extends Controller
 
     public function create(Business $business, Request $request)
     {
+        $businessUser = BusinessUserHelper::index($business['id'], Auth::user()['id']);
+        
+        if (!$businessUser) {
+            return abort(403);
+        } 
         $pemasok = $request['pemasok'] ? $request['pemasok'] : '';
         $kategori = $request['kategori'] ? $request['kategori'] : '';
         $brand = $request['brand'] ? $request['brand'] : '';
@@ -49,6 +61,11 @@ class StockController extends Controller
 
     public function store(Business $business, Request $request)
     {
+        $businessUser = BusinessUserHelper::index($business['id'], Auth::user()['id']);
+        
+        if (!$businessUser) {
+            return abort(403);
+        } 
         $create = Product::create([
             'pemasok' => strtoupper($request->pemasok),
             'brand' => strtoupper($request->brand),
@@ -74,6 +91,11 @@ class StockController extends Controller
 
     public function update(Business $business, Stock $stock, Request $request)
     {
+        $businessUser = BusinessUserHelper::index($business['id'], Auth::user()['id']);
+        
+        if (!$businessUser) {
+            return abort(403);
+        } 
         $stock->update([
             'satuan' => $request->satuan,
             'jumlah' => $request->jumlah,
@@ -97,6 +119,11 @@ class StockController extends Controller
 
     public function delete(Business $business, Stock $stock, Request $request)
     {
+        $businessUser = BusinessUserHelper::index($business['id'], Auth::user()['id']);
+        
+        if (!$businessUser) {
+            return abort(403);
+        } 
         $product = Product::find($stock['product_id'])->delete();
 
         if ($request->page == 'create') {
@@ -107,7 +134,7 @@ class StockController extends Controller
 
     public function detail(Business $business, Stock $stock)
     {
-
+        
         $response = [
             'message' => "Data Telah Tervalidasi",
             'status' => 'Success',
